requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
bert-score==0.3.13
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dash==2.16.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-holoniq-wordcloud==0.0.4
dash-html-components==2.0.0
dash-image-gallery==0.0.1
dash-table==5.0.0
dash_ag_grid==31.2.0
datasets==2.20.0
dill==0.3.8
evaluate==0.4.2
filelock==3.15.1
Flask==3.0.3
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.2.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==7.1.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
llvmlite==0.43.0
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numba==0.60.0
numpy==1.26.0
packaging==24.1
pandas==2.2.2
patsy==0.5.6
pillow==10.3.0
plotly==5.20.0
portalocker==2.8.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pynndescent==0.5.13
pyparsing==3.1.2
pyter3==0.3
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
retrying==1.3.4
rouge==1.0.1
sacrebleu==2.4.2
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
setuptools==70.0.0
six==1.16.0
statsmodels==0.14.2
sympy==1.12.1
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.4.1
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.2
tzdata==2024.1
umap-learn==0.5.6
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
wget==3.2
wordcloud==1.9.3
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
